many fixes and added features, notably better collection/model/view integration

+++ src/$collection.js
@@ -6,15 +6,16 @@
 (function() {
 
 	var collectionAPI = {
+
 		init: function() {
 			// make sure our values are unique across instances
 			this.items = [];
-			this.node = null;
-			this.view = null;
-			this.schema = null;
 		},
-		add: function(val) {
-//			console.log(val);
+		length: function() {
+			return this.items.length;
+		},
+		add: function(val, silent) {
+//			// console.log(val);
 			// supports batch add
 			var items = $isArray(val) ? val : [val];
 
@@ -24,7 +25,7 @@
 				if (that.view) {
 					// make sure we insert a view instance not a config obj
 					if ($isModel(item)) {
-//						console.log("adding view with model", item);
+//						// console.log("adding view with model", item);
 						// support overriding view model on view construction
 						item = that.view(item);
 //						console.log("makin a view", item);
@@ -32,38 +33,95 @@
 //						console.log("adding view", item);
 						item = $isView(item) ? item : that.view(item);
 					}
-				} else {
+				} else if  (that.schema) {
 					// make sure we insert a model instance not a config obj
-//					console.log("adding model", item);
-					item = $isModel(item) ? item : that.schema(item);
+					item = $isModel(item) ? item : that.schema.newInstance(item);
 				}
+
+				if ($isSpeaker(item)) {
+					that.listensTo(item);
+				}
+
 				that.items.push(item);
 			});
 
 			//resort regroup
-			if (this.sortIterator) {	this.items = $sortBy(this.items, this.sortIterator)}
-			if (this.groupByIterator) {	this.items = $groupBy(this.items, this.groupByIterator)}
-			if (this.isReversed) {		this.items.reverse()}
-			this.tell("change:add", items);
+			if (this.sortIterator) {	this.items = $sortBy(this.items, this.sortIterator);}
+			if (this.groupByIterator) {	this.items = $groupBy(this.items, this.groupByIterator);}
+			if (this.isReversed) {		this.items.reverse();}
+			if (!silent) {
+				this.tell("change:add", items);
+			}
 
 			return this;
 		},
 
-		sortBy: function(val) {
-//			console.log(this);
-			var iterator = $isFunction(val) ? val : function(obj) {
-//				console.log(obj, val);
-				return obj.model.get()[val];
+		push: function(item) {
+			this.add(item);
+			return this;
+		},
+
+		set: function(items) {
+			// todo drop views or models that are not in the new items collection?
+			this.items = [];
+			this.add(items, "silent");
+			this.tell("change:set", items);
+			return this;
+		},
+
+		// returns model instances
+		get: function(i) {
+			if ($isNumber(i)) {
+				return this.items[i];
+			} else if (i === "last") {
+				return this.items[this.items.length-1];
+			} else {
+				return this.items;
+			}
+		},
+
+		// returns model values
+		// todo: how does this relate to pluck?
+		getValues: function(i, key) {
+			var model, len = this.items.length;
+			if ($isNumber(i)) {
+				i = (i < 0) ? (len + i) % len : i;
+				model = this.items[i];
+			} else if (i === "last") {
+				model = this.items[len-1];
+			}
+
+			if (model) {
+				return $isFunction(val.get) ? model.get(key) : null;
+			} else {
+				return $map(this.items, function(val) {
+					return $isFunction(val.get) ? val.get(key) : val;
+				});
+			}
+		},
+
+		getGroupSortIterator: function(val) {
+			return function(obj) {
+				if ($isFunction(obj.get)) {
+					return obj.get()[val];
+				} else if ($isModel(obj.model)) {
+					return obj.model.get()[val];
+				} else {
+					return obj[val];
+				}
 			};
+		},
+
+		sortBy: function(val) {
+			var iterator = $isFunction(val) ? val : this.getGroupSortIterator(val);
 			this.sortIterator = iterator;
 			this.items = $sortBy(this.items, iterator);
-//			console.log(val, this.items);
 			this.tell("change:sortBy");
 			return this;
 		},
 
 		groupBy: function(val) {
-			var iterator = $isFunction(val) ? val : function(obj) { return obj.model.get()[val]; };
+			var iterator = $isFunction(val) ? val : this.getGroupSortIterator(val);
 			this.groupByIterator = iterator;
 			this.items = $groupBy(this.items, iterator);
 			this.tell("change:groupBy");
@@ -78,20 +136,29 @@
 		},
 
 		drop: function(obj) {
-			var i = this.items.indexOf(obj);
+			var i = obj ? this.items.indexOf(obj) : this.items.length-1;
 			if (i > -1) {
-				var old = this.items.splice(i, 1);
+				var old = this.items.splice(i, 1)[0];
 			}
-			this.tell("change:drop", old[0]);
-			this.view && old[0].drop();
+			this.tell("change:drop", old);
+			old && old.drop();
 			return this;
+		},
+
+		pop: function() {
+			var old = this.items.pop();
+			var snapshot = $isModel(old) ? old.get() : old ? old.model : null;
+			snapshot = $isModel(snapshot) ? snapshot.get() : old;
+			this.tell("change:pop", snapshot);
+			return snapshot;
 		},
 
 		dropAll: function() {
 			$map(this.items, function(item) {
 				// both views and models have a drop method
-				item.drop();
+				item.drop && item.drop();
 			});
+			this.items = [];
 			this.tell("change:dropAll");
 			return this;
 		},
@@ -101,7 +168,6 @@
 				this.node.innerHTML = "";
 				$each(this.items, function(view) {
 					if (view && $isElement(view.node)) {
-//						console.log(that.node.id);
 						this.node.appendChild(view.node);
 					}
 				}, this);
@@ -112,27 +178,36 @@
 
 	};
 
-	// used for view collections
+	var collectionBank = {};
 
 	// $collection function creates a collection instance
-	loot.extend("$collection", function(spec) {
+	function $collection(name, spec) {
 		var view = spec.view,
 			schema = spec.schema;
 
+		// console.log("collection extend", name, spec);
+
 		// return a new collection constructor
-		if (!schema && !view) {
-			throw new Error("$collection: Expected an object with a schema or view property");
+//		if (!schema && !view) {
+//			throw new Error("$collection: Expected an object with a schema or view property");
+//		}
+
+		if (name in collectionBank) {
+			// console.log(collectionBank, name);
+			throw new Error("collection name " + name + " already in use;");
 		}
 
 		var collection = $speak($extend($new(collectionAPI), spec));
 
 		// make sure we have a valid view constructor
 		if (view) {
+//			// console.log(view, $views(view));
 			view = $isString(view) ? $views(view) : view;
 			if ($isViewConstructor(view)) {
 				collection.view = view;
 				collection.draw = spec.draw ? spec.draw : collection.draw;
 			} else {
+//				// console.log(collection, view);
 				throw new Error("$collection: Invalid view type or constructor function");
 			}
 		} else if (schema) {
@@ -141,11 +216,49 @@
 			if ($isSchema(schema)) {
 				collection.schema = schema;
 			} else {
-				throw new Error("$collection: Invalid schema type or constructor function");
+				throw new Error("$collection: Invalid schema type or constructor function ");
 			}
 		}
 
+		if ($isString(collection.node)) {
+			collection.node = $el(collection.node);
+		}
+
+		if (collection.events && $isElement(collection.node)) {
+			$bindEventSpec(collection);
+		}
+
+		if ($isPlainObject(collection.listeners)) {
+			$bindListenerSpec(collection);
+		}
+
+		if ($isFunction(collection.init)) {
+			collection.init();
+		}
+
+		collectionBank[name] = collection;
+
 		return collection;
+	}
+
+
+	function $collections(name) {
+		if (!name) {
+			return collectionBank;
+		} else {
+			return collectionBank[name];
+		}
+	}
+
+	function $isCollection(obj) {
+		return !!(obj && obj.add && $isArray(obj.items) && obj.pop);
+	}
+
+	// expose methods
+	loot.extend({
+		$collection: $collection,
+		$collections: $collections,
+		$isCollection: $isCollection
 	});
 
 }());
